FWFemMeshで、PHFemMeshからGRMeshを作る機能を追加。 ロード後にロードしたオブジェクト(ルートのオブジェクトの子孫ノード) のLoaded()関数を呼ぶようにObjectに仕様を追加。
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@5085 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ include/Base/TMatrix.h
@@ -491,7 +491,7 @@ template <class AD, class XD, class BD>
 void solve(MatrixImp<AD>& a, VectorImp<XD>& x, const VectorImp<BD>& b, int* ip){
 	int i, j, ii;
 	TYPENAME XD::element_type t;
-	const size_t n = a.height();
+	const int n = (int)a.height();
 	for (i = 0; i < n; i++) {       // GaussÁ‹Ž–@‚ÌŽc‚è
 		ii = ip[i];  t = b[ii];
 		for (j = 0; j < i; j++) t -= a.item(ii, j) * x[j];
